Re-org emulated_hue and fix google home (#4708)
