+++ tests/components/emulated_hue/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for emulated_hue."""
